contextual_bandit_brain/requirements.txt
numpy>=1.23
matplotlib>=3.6
pytest>=7.0
PyYAML>=6.0
